Replace KvResponsiveImage with KvResponsiveImage2
